more YZViewCursor to its own files

+++ libyzis/swapfile.cpp
@@ -75,7 +75,7 @@ void YZSwapFile::flush() {
 void YZSwapFile::addToSwap( int inputs, int modifiers ) {
 	swapEntry e = { inputs, modifiers };
 	mHistory.append( e );
-	if ( mHistory.size() >= mParent->getLocalIntOption("updatecount") ) flush();
+	if ( mHistory.size() >= (int)mParent->getLocalIntOption("updatecount") ) flush();
 }
 
 void YZSwapFile::unlink() {
